+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/Navigation.kt
@@ -18,6 +18,7 @@ import co.electriccoin.zcash.ui.NavigationArguments.SEND_SCAN_RECIPIENT_ADDRESS
 import co.electriccoin.zcash.ui.NavigationTargets.ABOUT
 import co.electriccoin.zcash.ui.NavigationTargets.ADVANCED_SETTINGS
 import co.electriccoin.zcash.ui.NavigationTargets.CHOOSE_SERVER
+import co.electriccoin.zcash.ui.NavigationTargets.DELETE_WALLET
 import co.electriccoin.zcash.ui.NavigationTargets.EXPORT_PRIVATE_DATA
 import co.electriccoin.zcash.ui.NavigationTargets.HOME
 import co.electriccoin.zcash.ui.NavigationTargets.SCAN
@@ -35,6 +36,7 @@ import co.electriccoin.zcash.ui.design.animation.ScreenAnimation.popExitTransiti
 import co.electriccoin.zcash.ui.screen.about.WrapAbout
 import co.electriccoin.zcash.ui.screen.advancedsettings.WrapAdvancedSettings
 import co.electriccoin.zcash.ui.screen.chooseserver.WrapChooseServer
+import co.electriccoin.zcash.ui.screen.deletewallet.WrapDeleteWallet
 import co.electriccoin.zcash.ui.screen.exportdata.WrapExportPrivateData
 import co.electriccoin.zcash.ui.screen.home.WrapHome
 import co.electriccoin.zcash.ui.screen.scan.WrapScanValidator
@@ -136,6 +138,9 @@ internal fun MainActivity.Navigation() {
                 goChooseServer = {
                     navController.navigateJustOnce(CHOOSE_SERVER)
                 },
+                goDeleteWallet = {
+                    navController.navigateJustOnce(DELETE_WALLET)
+                },
             )
         }
         composable(CHOOSE_SERVER) {
@@ -158,6 +163,9 @@ internal fun MainActivity.Navigation() {
         composable(SUPPORT) {
             // Pop back stack won't be right if we deep link into support
             WrapSupport(goBack = { navController.popBackStackJustOnce(SUPPORT) })
+        }
+        composable(DELETE_WALLET) {
+            WrapDeleteWallet(goBack = { navController.popBackStackJustOnce(DELETE_WALLET) })
         }
         composable(ABOUT) {
             WrapAbout(goBack = { navController.popBackStackJustOnce(ABOUT) })
@@ -260,6 +268,7 @@ object NavigationArguments {
 object NavigationTargets {
     const val ABOUT = "about"
     const val ADVANCED_SETTINGS = "advanced_settings"
+    const val DELETE_WALLET = "delete_wallet"
     const val EXPORT_PRIVATE_DATA = "export_private_data"
     const val HOME = "home"
     const val CHOOSE_SERVER = "choose_server"